requirements.txt
# Media Processor Service Dependencies

# 任务队列
celery>=5.3.0
redis>=5.0.0
kombu>=5.3.0

# Web API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
sse-starlette>=1.6.0

# AI/ML
openai-whisper>=20231117
torch>=2.0.0
openai>=1.0.0

# 媒体处理
yt-dlp>=2023.12.0

# 工具
requests>=2.31.0
python-dotenv>=1.0.0
